Move parse_doc_title_w_placeholders to separate file
